requirements.txt
boto3==1.34.162
watchtower==3.1.0
python-dotenv==1.0.1
psutil==5.9.8
